# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/attachments/entity/{entityId}":{"get":{"description":"Retrieve all attachments associated with a specific entity","operationId":"AttachmentsController_findByEntityId","parameters":[{"name":"entityId","required":true,"in":"path","description":"The ID of the entity to get attachments for","schema":{"format":"uuid","example":"f47ac10b-58cc-4372-a567-0e02b2c3d479","type":"string"}}],"responses":{"200":{"description":"Attachments retrieved successfully","content":{"application/json":{"schema":{"type":"object","properties":{"attachments":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"fileUrl":{"type":"string"},"scope":{"type":"string"},"entityId":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}}}}}}}},"400":{"description":"Bad request - Invalid entity ID"},"500":{"description":"Internal server error"}},"summary":"Get attachments by entity ID","tags":["Attachments"]}},"/api/attachments":{"post":{"description":"Upload a file and create an attachment associated with an entity","operationId":"AttachmentsController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["file","name","scope","entityId"],"properties":{"file":{"type":"string","format":"binary","description":"File to upload"},"name":{"type":"string","description":"Name of the attachment","example":"receipt.pdf"},"scope":{"type":"string","enum":["TRANSACTION"],"description":"Scope of the attachment","example":"TRANSACTION"},"entityId":{"type":"string","format":"uuid","description":"ID of the entity associated with this attachment","example":"f47ac10b-58cc-4372-a567-0e02b2c3d479"}}}}}},"responses":{"201":{"description":"Attachment created successfully","content":{"application/json":{"schema":{"type":"object","properties":{"attachmentId":{"type":"string","format":"uuid","example":"a1b2c3d4-e5f6-7890-abcd-ef1234567890"}}}}}},"400":{"description":"Bad request - Invalid input data"},"500":{"description":"Internal server error"}},"summary":"Create attachment","tags":["Attachments"]}},"/api/auth/register":{"post":{"operationId":"AuthController_createRegister","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/refresh":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/transactions/categories":{"get":{"operationId":"TransactionsController_listCategories","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","type":"string","enum":["ASC","DESC"]}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Transactions"]},"post":{"operationId":"TransactionsController_createCategory","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTransactionCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["Transactions"]}},"/api/transactions/categories/{id}":{"put":{"operationId":"TransactionsController_updateCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTransactionCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["Transactions"]},"delete":{"operationId":"TransactionsController_deleteCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Transactions"]}},"/api/transactions/{id}":{"put":{"operationId":"TransactionsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTransactionDto"}}}},"responses":{"200":{"description":""}},"tags":["Transactions"]},"delete":{"operationId":"TransactionsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Transactions"]}},"/api/transactions":{"post":{"operationId":"TransactionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTransactionDto"}}}},"responses":{"201":{"description":""}},"tags":["Transactions"]},"get":{"operationId":"TransactionsController_list","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","type":"string","enum":["ASC","DESC"]}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Transactions"]}},"/api/transactions/summary/{period}":{"get":{"operationId":"TransactionsController_extractSummary","parameters":[{"name":"period","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Transactions"]}},"/api/payment-methods":{"post":{"operationId":"PaymentMethodController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePaymentMethodDto"}}}},"responses":{"201":{"description":"Payment method created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentMethodDto"}}}}},"summary":"Create a new payment method","tags":["Payment Methods"]},"get":{"operationId":"PaymentMethodController_list","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","type":"string","enum":["ASC","DESC"]}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}}],"responses":{"200":{"description":"List of payment methods"}},"summary":"List all payment methods","tags":["Payment Methods"]}},"/api/payment-methods/{id}":{"delete":{"operationId":"PaymentMethodController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Payment method deleted successfully"}},"summary":"Delete a payment method","tags":["Payment Methods"]}}},"info":{"title":"Base APi","description":"Base Api","version":"1.0","contact":{}},"tags":[{"name":"Base Api","description":""}],"servers":[{"url":"http://localhost:3000/","description":"DEV"}],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{}},"Credentials":{"type":"object","properties":{}},"CreateTransactionCategoryDto":{"type":"object","properties":{"name":{"type":"string","example":"Personal Expenses","description":"Name of the transaction category"},"description":{"type":"object","example":"Expenses related to personal items and services","description":"Description of the transaction category","nullable":true},"types":{"type":"array","example":["COMMON","AGENCY","LINE"],"description":"Types of category","items":{"type":"string","enum":["COMMON","AGENCY","LINE"]}}},"required":["name","description","types"]},"UpdateTransactionCategoryDto":{"type":"object","properties":{"name":{"type":"string","example":"Personal Expenses","description":"Name of the transaction category"},"description":{"type":"object","example":"Expenses related to personal items and services","description":"Description of the transaction category","nullable":true}},"required":["name","description"]},"UpdateTransactionDto":{"type":"object","properties":{"categoryId":{"type":"string","description":"Category ID of this transaction","example":"123e4567-e89b-12d3-a456-426614174002"},"description":{"type":"string","description":"Transaction description","minLength":3,"example":"Purchase at supermarket"},"paymentMethodId":{"type":"object","description":"Payment Method ID of this transaction","example":"123e4567-e89b-12d3-a456-426614174003","nullable":true},"type":{"type":"string","description":"Transaction type","enum":["INCOME","EXPENSE"],"example":"EXPENSE"},"amount":{"type":"number","description":"Transaction amount in currency","example":150.5,"nullable":true}}},"CreateTransactionDto":{"type":"object","properties":{"categoryId":{"type":"string","description":"Category ID of this transaction","example":"123e4567-e89b-12d3-a456-426614174002"},"description":{"type":"string","description":"Transaction description","minLength":3,"example":"Purchase at supermarket"},"paymentMethodId":{"type":"object","description":"Payment Method ID of this transaction","example":"123e4567-e89b-12d3-a456-426614174003","nullable":true},"type":{"type":"string","description":"Transaction type","enum":["INCOME","EXPENSE"],"example":"EXPENSE"},"amount":{"type":"number","description":"Transaction amount in currency","example":150.5,"nullable":true},"createdAt":{"type":"object","description":"Transaction creation date","example":"2024-01-07T10:30:00Z","examples":["2024-01-07T10:30:00Z","2024-06-15T08:20:00Z","2024-12-31T23:59:59Z",null],"nullable":true}},"required":["categoryId","description","paymentMethodId","type"]},"CreatePaymentMethodDto":{"type":"object","properties":{}},"PaymentMethodDto":{"type":"object","properties":{}}}}}